+++ src/extension.ts
@@ -3,6 +3,7 @@ import { MarkerTreeViewProvider } from './stackView';
 import { extensionEnv } from './extensionEnv';
 import { registerCommands } from './commands';
 import { output } from './output';
+import { markerService } from './markerService';
 
 export function activate(context: vscode.ExtensionContext) {
   extensionEnv.setExtensionContext(context);
@@ -13,6 +14,9 @@ export function activate(context: vscode.ExtensionContext) {
   output.activate();
 
   output.log('Code Explorer is activated');
+
+  context.subscriptions.push(output);
+  context.subscriptions.push(markerService);
 }
 
 export function deactivate() {}